+++ src/index.ts
@@ -1,5 +1,0 @@
-//
-// The app entrypoint
-//
-
-export * from './types'