+++ src/module.ts
@@ -1,0 +1,11 @@
+//
+// The app entrypoint
+//
+
+export * from './attendance'
+export * from './carbon'
+export * from './conference'
+export * from './interpret'
+export * from './lib'
+export * from './registration'
+export * from './utils'